refactored ContentPersistor - now has abstract ContentPersistorBase that can be used for other persistence implementations
